// src/types/marketplace-api.ts
export interface components {
    schemas: {
// ...
        TransactionPurchaseDetails: {
            /**
             * Format: date
             * @description The transaction date, payment date or due date (for transactions having Net 14/30 payment terms)
             */
            saleDate: string;
            /**
             * @description The unique identifier for this license type.
             * @enum {string}
             */
            licenseType: "ACADEMIC" | "COMMERCIAL" | "COMMUNITY" | "EVALUATION" | "OPEN_SOURCE";
            /**
             * @description The platform on which the app runs.
             * @enum {string}
             */
            hosting: "Server" | "Data Center" | "Cloud";
            /**
             * @description The time period for which the transaction is billed.
             * @enum {string}
             */
            billingPeriod: "Monthly" | "Annual";
            /**
             * @description Shows whether there is change in the billing period of an app from the (n-1)th transaction to nth transaction.
             * @enum {string}
             */
            changeInBillingPeriod?: "Yes" | "No";
            /**
             * @description The (n-1)th transaction’s billing period of an app.
             * @enum {string}
             */
            oldBillingPeriod?: "Monthly" | "Annual" | "Other";
            /** @description The user/edition count. For example 100 Users, Unlimited Users, Per Unit Pricing (593 Users) */
            tier: string;
            /**
             * @description Shows whether there is change in the user/edition count from the (n-1)th transaction to nth transaction
             * @enum {string}
             */
            changeInTier?: "Increase" | "Decrease" | "Not Applicable";
            /** @description The (n-1)th transaction’s user/edition count.  For example 100 Users, Unlimited Users, Per Unit Pricing (593 Users) */
            oldTier?: string;
            /**
             * @description Name of the parent product on which the app is hosted. Applies to cloud licenses only.
             * @enum {string}
             */
            parentProductName?: "Jira" | "Confluence" | "Pending";
            /**
             * @description A short description of the pricing tier or user level of the license parent product. Applies to cloud licenses only.
             * @enum {string}
             */
            parentProductEdition?: "Jira" | "Confluence" | "Pending";
            /**
             * @description Shows whether there is change in the parent edition from the (n-1)th transaction to nth transaction. Applies to cloud licenses only
             * @enum {string}
             */
            changeInParentProductEdition?: "Yes" | "No";
            /** @description The (n-1)th transaction’s parent edition. Applies to cloud licenses only */
            oldParentProductEdition?: string;
            /**
             * Format: float
             * @description The sale price in USD, for example 100.0
             */
            purchasePrice: number;
            /**
             * Format: float
             * @description The portion of the sale price which will be paid to the partner in USD, for example 90.0
             */
            vendorAmount: number;
            /**
             * Format: float
             * @description For partner sales, the amount discounted before the sale price.
             */
            partnerDiscountAmount?: number;
            /** @description List of discounts that are applied on the transaction. */
            discounts?: components["schemas"]["TransactionDiscount"][];
            /** @description The reason of refund for the transaction, for example 'REFUND_CANCELLED_PURCHASE' */
            refundReason?: string;
            /** @description The reason of credit note for the transaction, for example '30 DAY REFUND REQUEST' */
            creditNoteReason?: string;
            /**
             * @description The display name of the sale type.
             * @enum {string}
             */
            saleType: "New" | "Refund" | "Renewal" | "Upgrade";
            /**
             * Format: date
             * @description The date on which license maintenance begins.
             */
            maintenanceStartDate: string;
            /**
             * Format: date
             * @description The date on which license maintenance ends.
             */
            maintenanceEndDate: string;
            /**
             * @description Credit limit offered to customers.
             * @enum {string}
             */
            paymentTerms?: "Net14" | "Net30";
            originalTransactionDetails?: components["schemas"]["OriginalTransactionDetails"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}